netpacket/if_addr: add interface address header
